Allow user to overwrite the |host| header while generating the request
As we discussed in #255, it's possible for us to provide
some way to sepcifiy the generated request's |host| header.

Now when the incoming request stuct contains a |host| header,
we will use it's value as outgoing |host| header's value.

If there's no |host| header specified, we still use the authority
part of uri as |host| header.

This behavior is complied with RFC 6455 and close issue #255